Rename clock module to clock_time
